admin/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
bcrypt==4.1.2
Pillow==10.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
